unit test for surcharge rate case is added to insurance service tests, surcharge rate is applied to products on insurance service

+++ src/Insurance.Api/Models/Request/UpdateSurchargeRateRequest.cs
@@ -3,7 +3,7 @@
     public class UpdateSurchargeRateRequest
     {
         public string Name { get; set; }
-        public double Rate { get; set; }
+        public int Rate { get; set; }
         public int ProductTypeId { get; set; }
     }
 }
